refactor: agrupamento de classes
