+++ ManualDeManutencao.java
@@ -1,5 +1,3 @@
-package WegOne;
-
 import java.util.Date;
 import java.util.Scanner;
